% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{augmented_diag}
\alias{augmented_diag}
\title{Returns an offset diagonal matrix.}
\usage{
augmented_diag(d, offset)
}
\arguments{
\item{d}{Matrix dimensional (square matrix)}

\item{offset}{Offset from the diagonal}
}
\value{
A matrix with an offset diagonal of ones.
}
\description{
Returns an offset diagonal matrix.
}
\examples{
d <- 10
offset <- 1
augmented_diag(d, offset) # zeroes everywhere except 1s on the +1 diag
}
