% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AugmentedDiag}
\alias{AugmentedDiag}
\title{Returns an offset diagonal matrix.}
\usage{
AugmentedDiag(d, offset)
}
\arguments{
\item{d}{Matrix dimensional (square matrix)}

\item{offset}{Offset from the diagonal}
}
\value{
A matrix with an offset diagonal of ones.
}
\description{
Returns an offset diagonal matrix.
}
