% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{check_thetas}
\alias{check_thetas}
\title{Checks that \code{|theta_1| <= theta_2}.}
\usage{
check_thetas(theta_1, theta_2)
}
\arguments{
\item{theta_1}{Off-diagonal value or network effect.}

\item{theta_2}{Diagonal value or momentum effect.}
}
\value{
An error if \code{|theta_1| > theta_2}, nothing happens otherwise.
}
\description{
Checks that \code{|theta_1| <= theta_2}.
}
