% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{CheckThetas}
\alias{CheckThetas}
\title{Checks that \code{|theta_1| < theta_2}.}
\usage{
CheckThetas(theta_1, theta_2)
}
\arguments{
\item{theta_1}{Off-diagonal value or network effect.}

\item{theta_2}{Diagonal value or momentum effect.}
}
\description{
Checks that \code{|theta_1| < theta_2}.
}
