% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CleanData}
\alias{CleanData}
\title{Clean data using STL / Loess function.}
\usage{
CleanData(data, frequency = 24, s.window = 24, t.window = 24, ...)
}
\arguments{
\item{data}{Data to clean.}

\item{frequency}{Time series frequency of observations}

\item{s.window}{Seasonal window.}

\item{t.window}{Trend window.}

\item{...}{additional inputs given to stats::stl}
}
\value{
List with the cleaned time series, the remainders of the cleaning and their standard deviations.
}
\description{
Clean data using STL / Loess function.
}
