% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_col}
\alias{concatenate_col}
\title{Concatenate \code{col_vec} columnwise \code{n} times}
\usage{
concatenate_col(col_vec, n)
}
\arguments{
\item{col_vec}{a vector to copy/concatenate}

\item{n}{number of times to concatenate (number of columns)}
}
\value{
A matrix where \code{col_vec} is concatenated
\code{n} times along the column axis.
}
\description{
Concatenate \code{col_vec} columnwise \code{n} times
}
\examples{
vec_to_concat <- seq(10)
n <- 5
concatenate_col(vec_to_concat, n) # (10, 5) matrix created
}
