% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConcatenateCol}
\alias{ConcatenateCol}
\title{Concatenate \code{col_vec} columnwise \code{n} times}
\usage{
ConcatenateCol(col_vec, n)
}
\arguments{
\item{col_vec}{a vector to copy/concatenate}

\item{n}{number of times to concatenate (number of columns)}
}
\value{
A matrix where \code{col_vec} is concatenated \code{n} times along the column axis.
}
\description{
Concatenate \code{col_vec} columnwise \code{n} times
}
