% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grou.R
\name{core_node_mle}
\alias{core_node_mle}
\title{Computes the MLE numerator and denominator of a batch of data}
\usage{
core_node_mle(times, data, thresholds = NA)
}
\arguments{
\item{times}{Times at which data is given}

\item{data}{Values to compute the MLE with.}

\item{thresholds}{Jump threshold values. Default is \code{NA} (no filtering).}
}
\value{
A list of two matrices (numerator and denominator).
}
\description{
Computes the MLE numerator and denominator of a batch of data
}
