% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularisation.R
\name{finding_lambda}
\alias{finding_lambda}
\title{Finding \code{lambda} for regularisation.}
\usage{
finding_lambda(
  times,
  data,
  thresholds,
  reg = "l1",
  div = 1e+05,
  gamma = NA,
  use_scaling = F,
  cross = F
)
}
\arguments{
\item{times}{Times at which data is given}

\item{data}{Values to compute the MLE with.}

\item{thresholds}{Jump threshold values.}

\item{reg}{Type of penalty (\code{l1}, \code{l2} or \code{adaptive}).}

\item{div}{Batch size/divisor to avoid large memory allocation.}

\item{gamma}{Adaptive MLE scaling parameter.}

\item{use_scaling}{Brownian motion covariance matrix scaling
in the likelihood.}

\item{cross}{Cross-validation or not (defaults to \code{FALSE}).}
}
\value{
List of top-choice for lambda, the collection of lambdas test
and corresponding log-likelihood values
}
\description{
Finding \code{lambda} for regularisation.
}
