% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{isolated_network}
\alias{isolated_network}
\title{Creates adjacency matrix with max degree zero.}
\usage{
isolated_network(d, theta_1, theta_2)
}
\arguments{
\item{d}{Number of nodes.}

\item{theta_1}{Off-diagonal value or network effect.}

\item{theta_2}{Diagonal value or momentum effect.}
}
\value{
adjacency matrix of a graph with max degree zero.
}
\description{
Creates adjacency matrix with max degree zero.
}
