% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{IsolatedNetwork}
\alias{IsolatedNetwork}
\title{Creates adjacency matrix with max degree zewro.}
\usage{
IsolatedNetwork(d, theta_1, theta_2)
}
\arguments{
\item{d}{Number of nodes.}

\item{theta_1}{Off-diagonal value or network effect.}

\item{theta_2}{Diagonal value or momentum effect.}
}
\value{
adjacency matrix of a graph with max degree zero.
}
\description{
Creates adjacency matrix with max degree zewro.
}
