% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grou.R
\name{node_mle}
\alias{node_mle}
\title{Computes the psi-GrOU MLE of a batch of data}
\usage{
node_mle(times, data, output = "vector")
}
\arguments{
\item{times}{Times at which data is given}

\item{data}{Values to compute the MLE with.}

\item{output}{Output type: either "vector"or "matrix".}
}
\value{
the psi-GrOU (node-wise) MLE
}
\description{
Computes the psi-GrOU MLE of a batch of data
}
\examples{
n <- 1000
d <- 10
times <- seq(n)
delta_time <- 0.01
noise <- matrix(rnorm(n * d, sd = sqrt(delta_time)), ncol = d)
data <- construct_path(
  diag(d),
  noise = noise, y_init = rep(0, d), delta_time = delta_time
)
node_mle(times, data)
}
