% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grou.R
\name{node_mle_components}
\alias{node_mle_components}
\title{Returns the components (i.e. numerator and denominator) of the MLE}
\usage{
node_mle_components(times, data, thresholds, div = 1e+05, output = "vector")
}
\description{
Returns the components (i.e. numerator and denominator) of the MLE
}
