% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grou.R
\name{node_mle_long}
\alias{node_mle_long}
\title{Constructs the Node MLE with jump thresholding}
\usage{
node_mle_long(times, data, thresholds, div = 1e+05, output = "vector")
}
\arguments{
\item{times}{Times at which data is given}

\item{data}{Values to compute the MLE with.}

\item{thresholds}{Jump threshold values.}

\item{div}{Batch size/divisor to avoid large memory allocation.}

\item{output}{String to indicate
the form of the output: \code{vector} or \code{matrix}.}
}
\value{
The GrOU MLE in matrix/vector form in psi parametrisation.
}
\description{
Constructs the Node MLE with jump thresholding
}
