% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{lattice_network}
\alias{lattice_network}
\title{Lattice/grid-like network with max degree of four across all nodes.}
\usage{
lattice_network(d, theta_1, theta_2, directed = F)
}
\arguments{
\item{d}{Number of nodes.}

\item{theta_1}{Off-diagonal value or network effect.}

\item{theta_2}{Diagonal value or momentum effect.}

\item{directed}{Boolean if lower triangular is the opposite
to the upper triangular matrix.}
}
\description{
Lattice/grid-like network with max degree of four across all nodes.
}
